Creating component post-detail
